Added 3 pins for potentiometer
Added another 3 pin connector for the potentiometer to be wired to in the controller/transmitter side.  Currently both signals (transmitter or receiver) are on the same line but only one will ever be used at a given time anyway so shouldn't pose any issue.  Redid the routing still seems pretty clean.  Time to make a box for this thing.